Derive RelaySMSException from RelayAPIException
